Add sql query for finding cheeps
Allows embedded resources such that sql statements can be seperated from the source code. Also added query for finding all cheeps using parameters, which will have to be replaced and validated by parameterisation in source code. It includes a LIKE comparison such that cheeps can be filtered for a specific author, which can optionally be replaced by the wildcard string to allow all authors. A limit and offset are added for paginate the query, where limit will be used for limiting the number of cheeps on a page and the offset will be used for finding the page as a multiple of the limit parameter.