Refactor SQL queries to be embedded resources
SQLite hardcoded into the source code is bad practice as stated in lecture 05 as the data engineer and the software developer might be different people with different areas of expertise. Furthermore it is good for separation of concerns and helps with readibility. I have made it so the SQL queries are read from embedded resources with the sql queries saved as readonly strings with parameter placeholders.