refactor database design to use EF Core
In changing our database design to be using the Entity Framework instead, we have created data models for the author and cheeps and created a database context object, however this also meant we had to remove many methods and comment out a bunch of method bodies such that we could still compile. We have made it so queries previously in the DBFacade (now the new ChirpCbContext) throws an error when trying to use them, but this needs to be addressed in a later commit before merging into main.

Co-Authored-By: Maryna Lvova <228911840+alexmaryna@users.noreply.github.com>