WIP:buggy:Implement CheepRepository
We have implemented the requirements of the issue #52 however in order not to do the work of the other issues ( #51 and #56 ) which are assigned to other group members we had to leave the methods throwing not implemented exceptions. Furthermore we are unsure if the CheepService is to be replaced by the CheepRepository as a lot of the functionality seems to need to be moved to the latter.

Also we fixed some bugs of our previous commit.

Co-Authored-By: Maryna Lvova <228911840+alexmaryna@users.noreply.github.com>