Add login functionality
Added login functionality in shared layout pages. For both cases in which user is logged in or not, displays different items in navigation bar.

The pages for Register, Login, Logout are not created but are linked to, as creating these pages are part of 1.a) tasks. When these pages are created via using the ASP.NET core identity, the subsequent page links are ready for use.