Fix CheepRepository + Service UnitTests
After the Identity refactor, these tests were not fixed. Also one of the tests was wrong in that it used the DBInitialiser to seed the test database, however that seeder does not contain the author "Alice", as is tested against. This was probably meant to test against the SeedDatabase static method present in the test class instead.