Add DeleteAuthor method
Now an author can be deleted from every reference in the database. Only missing an implementation with a button in the future User Information page.

I renamed the files testing repository and service methods as we have been adding tests for both Cheeps and Authors in the same file.

I have not tested the DeleteAuthor in AuthorService, as it is simply calling a method of the same name in the AuthorRepository.