Add return value on DeleteAuthor
Adding a return value on DeleteAuthor, allows us to use the Author after it has been deleted for display purposes, showing on the future deleting page confirmation that the user X was deleted.

Since I modified AuthorService's DeleteAuthor to return a DTO I also added tests for this.

I think this method of testing in the Service as shown in commit, is a bit cleaner as it is only the database context that is mocked and we can then use real repository and service objects on this database context to confirm methods run as they should.